Java: The core count is an int, not a double

+++ src/bindings/java/org/simgrid/s4u/Host.java
@@ -28,7 +28,7 @@ public class Host {
   public double get_pstate_speed(int pstate) { return simgridJNI.Host_get_pstate_speed(swigCPtr, this, pstate); }
   public double get_load() { return simgridJNI.Host_get_load(swigCPtr); }
 
-  public double get_core_count() { return simgridJNI.Host_get_core_count(swigCPtr); }
+  public int get_core_count() { return simgridJNI.Host_get_core_count(swigCPtr); }
 
   public boolean is_on() { return simgridJNI.Host_is_on(swigCPtr, this); }
   public void turn_off() { simgridJNI.Host_turn_off(swigCPtr); }
